% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{dap_on}
\alias{dap_on}
\title{Check if the MICADO DAP server is running}
\usage{
dap_on()
}
\value{
Logical
}
\description{
For testing purposes, one can check whether the background R process that is
still serving the HTTP API is still active.
}
\seealso{
\code{\link[=dap_start]{dap_start()}}, \code{\link[=dap_stop]{dap_stop()}}, \code{\link[=dap_url]{dap_url()}}, \code{\link[=dap_request]{dap_request()}}
}
