% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{dap_request}
\alias{dap_request}
\title{Submit an HTTP request to the MICADO DAP}
\usage{
dap_request(
  gamma = NULL,
  pncc = NULL,
  ancc = NULL,
  case = NULL,
  legacy = FALSE,
  ip = "localhost"
)
}
\arguments{
\item{gamma}{Data frame with the GAMMA measurement data, including columns
nuclide, energy, time, detector, net and background. See \code{\link{vc1_gamma}} for
an example.}

\item{pncc}{PNCC reals per second. See \code{\link{vc1_pncc}} for an example.}

\item{ancc}{ANCC counts per second. See \code{\link{vc1_ancc}} for an example.}

\item{case}{The virtual case to consider. Either \code{"vc1"} or \code{"vc3"}.}

\item{legacy}{Logical flag to indicate whether the Plutonium isotopic vector
should be considered unknown or not (defaults to \code{FALSE}).}

\item{ip}{IP address of the machine that is running the MICADO DAP server.
Can be \code{"localhost"} (default) for local testing purposes.}
}
\value{
The probabilistic interpretation results, as a list with a \code{masses}
data frame and a \code{diagnostics} list.
}
\description{
Submit an HTTP request to the MICADO DAP
}
\seealso{
\code{\link[=dap_start]{dap_start()}}, \code{\link[=dap_stop]{dap_stop()}}, \code{\link[=dap_on]{dap_on()}}, \code{\link[=dap_url]{dap_url()}}
}
