% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{dap_start}
\alias{dap_start}
\title{Start the MICADO DAP server}
\usage{
dap_start(cores = 1)
}
\arguments{
\item{cores}{The number of cores to use for the probabilistic models. Can be
1 (default), 2 or 4.}
}
\description{
For accessing the MICADO DAP through the HTTP API, the server has to be
started from R first.
}
\seealso{
\code{\link[=dap_stop]{dap_stop()}}, \code{\link[=dap_on]{dap_on()}}, \code{\link[=dap_url]{dap_url()}}, \code{\link[=dap_request]{dap_request()}}
}
