% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{dap_stop}
\alias{dap_stop}
\title{Stop the MICADO DAP server}
\usage{
dap_stop()
}
\description{
The MICADO DAP server can always be stopped explicitly if needed. Normally,
however, the background R process will be killed whenever R exits.
}
\seealso{
\code{\link[=dap_start]{dap_start()}}, \code{\link[=dap_on]{dap_on()}}, \code{\link[=dap_url]{dap_url()}}, \code{\link[=dap_request]{dap_request()}}
}
