% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{dap_url}
\alias{dap_url}
\title{Get the URL at which the HTTP API can be accessed}
\usage{
dap_url()
}
\value{
The MICADO DAP URL
}
\description{
This is useful to check where to access the HTTP API from a remote machine.
}
\seealso{
\code{\link[=dap_start]{dap_start()}}, \code{\link[=dap_stop]{dap_stop()}}, \code{\link[=dap_on]{dap_on()}}, \code{\link[=dap_request]{dap_request()}}, \code{\link[=ip]{ip()}}
}
