% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_fit}
\alias{get_fit}
\title{Access the underlying CmdStanMCMC object}
\usage{
get_fit(model_result)
}
\arguments{
\item{model_result}{Output of a call to \code{\link[=vc1_model]{vc1_model()}} or \code{\link[=vc3_model]{vc3_model()}}}
}
\value{
The underlying \code{CmdStanMCMC} object
}
\description{
Access the underlying CmdStanMCMC object
}
\seealso{
\code{\link[=vc1_model]{vc1_model()}}, \code{\link[=vc3_model]{vc3_model()}}, \code{\link[=get_json]{get_json()}}
}
