% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_json}
\alias{get_json}
\title{Access the JSON for DigiWaste Bridge}
\usage{
get_json(model_result)
}
\arguments{
\item{model_result}{Output of a call to \code{\link[=vc1_model]{vc1_model()}} or \code{\link[=vc3_model]{vc3_model()}}}
}
\value{
The JSON for DigiWaste Bridge
}
\description{
Access the JSON for DigiWaste Bridge
}
\seealso{
\code{\link[=vc1_model]{vc1_model()}}, \code{\link[=vc3_model]{vc3_model()}}, \code{\link[=get_fit]{get_fit()}}
}
