% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-api.R
\name{ip}
\alias{ip}
\title{Find a machine's IP address through system commands.}
\usage{
ip(machine = NULL)
}
\arguments{
\item{machine}{Name of the machine to get the IP address for. If \code{NULL}
(default), the current machine is used.}
}
\value{
An IP address.
}
\description{
For the current machine, we use the first IPv4 listed by \code{ipconfig}, for
a remote machine, the address returned by \code{nslookup} is used.
}
\seealso{
\code{\link[=dap_url]{dap_url()}}
}
