% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vc1_ancc}
\alias{vc1_ancc}
\title{VC1 ANCC measurement data}
\format{
Length-one numeric vector with ANCC counts per second.
}
\source{
MICADO Consortium.
}
\usage{
vc1_ancc
}
\description{
Simulated data based on the "true" MCNP model.
}
\keyword{datasets}
