% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vc1_gamma}
\alias{vc1_gamma}
\title{VC1 Gamma measurement data}
\format{
A data frame with 85 rows and 6 columns:
\describe{
\item{nuclide}{Radionuclide name}
\item{energy}{Gamma peak energy}
\item{time}{Measurement time}
\item{detector}{Detector ID}
\item{net}{Net count}
\item{background}{Background count}
}
}
\source{
MICADO Consortium.
}
\usage{
vc1_gamma
}
\description{
Simulated data based on the "true" MCNP model.
}
\keyword{datasets}
