% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{vc3_model}
\alias{vc3_model}
\title{Probabilistic model for Virtual Case 3}
\usage{
vc3_model(gamma = NULL, pncc = NULL, legacy = FALSE, cores = 1, quiet = TRUE)
}
\arguments{
\item{gamma}{Data frame with the GAMMA measurement data, including columns
nuclide, energy, time, detector, net and background. See \code{\link{vc1_gamma}} for
an example.}

\item{pncc}{PNCC reals per second. See \code{\link{vc1_pncc}} for an example.}

\item{legacy}{Logical flag to indicate whether the Plutonium isotopic vector
should be considered unknown or not (defaults to \code{FALSE}).}

\item{cores}{The number of cores to use for the probabilistic models. Can be
1 (default), 2 or 4.}

\item{quiet}{Whether to suppress the Stan messages. Defaults to \code{TRUE}.}
}
\value{
A list with a \code{masses} data frame, and a \code{diagnostics} list.
}
\description{
Probabilistic model for Virtual Case 3
}
\seealso{
\code{\link[=vc1_model]{vc1_model()}}, \code{\link[=get_fit]{get_fit()}}, \code{\link[=get_json]{get_json()}}
}
