% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vc3_pncc}
\alias{vc3_pncc}
\title{VC3 PNCC measurement data}
\format{
Length-one numeric vector with PNCC reals per second.
}
\source{
MICADO Consortium.
}
\usage{
vc3_pncc
}
\description{
Simulated data based on the "true" MCNP model.
}
\keyword{datasets}
